Stnadardize code
